requirements.txt
python-dotenv==1.1.0
langchain-core==0.3.58
langchain-openai==0.3.16
langgraph==0.4.1
langsmith
firecrawl==2.5.3
openai==1.77.0
pytest==8.3.5
pytest-asyncio==0.26.0
PyMuPDF==1.24.8
camelot-py==0.11.0
tomli
pyyaml
tenacity
firecrawl-py
google-api-python-client==2.169
toml
supabase>=2.10.0
python-dotenv>=1.0.1 